@@ -20,6 +20,6 @@
   },
   "devDependencies": {
     "bob-the-bundler": "^1.4.1",
-    "typescript": "^4.3.4"
+    "typescript": "^4.8.3"
   }
 }
