@@ -5,16 +5,16 @@
 
 'use strict'
 
-import { afterAll, beforeAll, expect, test } from '@jest/globals'
-const {
+import { expect, test } from '@jest/globals'
+import {
   GraphQLSchema,
   GraphQLObjectType,
   GraphQLString,
   graphql
-} = require('graphql')
+} from 'graphql';
+import type { OpenAPIV3 } from 'openapi-types';
 
 import * as Oas3Tools from '../src/oas_3_tools'
-import { PathItemObject } from '../src/types/oas3'
 
 test('Applying sanitize multiple times does not change outcome', () => {
   const str = 'this Super*annoying-string()'
@@ -145,7 +145,6 @@ test('Properly treat null values during sanitization', () => {
       name: 'Query',
       fields: {
         User: {
-          name: 'name',
           type: new GraphQLObjectType({
             name: 'user',
             fields: {
@@ -171,7 +170,7 @@ test('Properly treat null values during sanitization', () => {
     }
   }`
 
-  graphql(schema, query).then((result) => {
+  graphql({schema, source: query}).then((result) => {
     expect(result).toEqual({
       data: {
         User: {
@@ -202,17 +201,9 @@ test('Handle encoded JSON pointer references', () => {
     Oas3Tools.resolveRef('#/paths/~1users~1%7bid%7d/description', oas)
   ).toBe('one')
 
-  function getPathItemObject(description): PathItemObject {
+  function getPathItemObject(description): OpenAPIV3.PathItemObject {
     return {
       description,
-      get: {},
-      put: {},
-      post: {},
-      delete: {},
-      options: {},
-      head: {},
-      patch: {},
-      trace: {}
     }
   }
 })
