@@ -4,10 +4,9 @@
 // License text available at https://opensource.org/licenses/MIT
 
 // Type imports:
-import { GraphQLOperationType, SubscriptionContext } from './graphql'
-import { GraphQLFieldResolver, GraphQLResolveInfo } from 'graphql'
-import crossFetch from 'cross-fetch'
-import FormData from 'form-data'
+import { fetch, FormDataLimits} from '@whatwg-node/fetch';
+import { GraphQLFieldResolver, GraphQLResolveInfo, OperationTypeNode } from 'graphql'
+import { SubscriptionContext } from './graphql'
 
 /**
  * Type definition of the options that users can pass to OpenAPI-to-GraphQL.
@@ -86,7 +85,7 @@ export type RequestOptions<TSource, TContext, TArgs> = Omit<
  *
  * Based on: https://github.com/form-data/form-data#custom-options
  */
-export type FileUploadOptions = ConstructorParameters<typeof FormData>[0]
+export type FileUploadOptions = FormDataLimits
 
 export type Options<TSource, TContext, TArgs> = Partial<
   InternalOptions<TSource, TContext, TArgs>
@@ -155,7 +154,7 @@ export type InternalOptions<TSource, TContext, TArgs> = {
    * The field is identifed first by the title of the OAS, then the path of the
    * operation, and lastly the method of the operation.
    */
-  selectQueryOrMutationField?: OasTitlePathMethodObject<GraphQLOperationType>
+  selectQueryOrMutationField?: OasTitlePathMethodObject<OperationTypeNode>
 
   /**
    * Sets argument name for the payload of a mutation to 'requestBody'
@@ -362,5 +361,5 @@ export type InternalOptions<TSource, TContext, TArgs> = {
   /**
    * Custom W3 Compatible `fetch` implementation
    */
-  fetch: typeof crossFetch
+  fetch: typeof fetch
 }
