@@ -5,16 +5,22 @@
 
 'use strict'
 
-import { graphql, GraphQLObjectType, GraphQLSchema } from 'graphql'
 import { afterAll, beforeAll, expect, test } from '@jest/globals'
+import { readFileSync } from 'fs'
+import { graphql, GraphQLObjectType, GraphQLSchema } from 'graphql'
+import { join } from 'path'
 
 import * as openAPIToGraphQL from '../src/index'
 import { startServer, stopServer } from './example_api2_server'
 
-const oas = require('./fixtures/example_oas2.json')
 const PORT = 3004
-// Update PORT for this test case:
-oas.servers[0].variables.port.default = String(PORT)
+function getOas() {
+  const oasStr = readFileSync(join(__dirname, './fixtures/example_oas2.json'), 'utf8');
+  const oas = JSON.parse(oasStr);
+  // update PORT for this test case:
+  oas.servers[0].variables.port.default = String(PORT);
+  return oas;
+}
 
 let createdSchema: GraphQLSchema
 
@@ -30,7 +36,7 @@ let createdSchema: GraphQLSchema
 beforeAll(() => {
   return Promise.all([
     openAPIToGraphQL
-      .createGraphQLSchema(oas, { operationIdFieldNames: true })
+      .createGraphQLSchema(getOas(), { operationIdFieldNames: true })
       .then(({ schema, report }) => {
         createdSchema = schema
       }),
@@ -51,6 +57,7 @@ afterAll(() => {
  * autogenerated field name based on the path, i.e. getUser
  */
 test('The option operationIdFieldNames should allow both operations to be present', () => {
+  const oas = getOas();
   let oasGetCount = 0
   for (let path in oas.paths) {
     for (let method in oas.paths[path]) {
@@ -64,20 +71,20 @@ test('The option operationIdFieldNames should allow both operations to be presen
 
 test('Querying the two operations', () => {
   const query = `query {
-    getUser {
-      name
-    }
     user {
       name
     }
+    User {
+      name
+    }
   }`
-  return graphql(createdSchema, query).then((result) => {
+  return graphql({schema: createdSchema, source: query}).then((result) => {
     expect(result).toEqual({
       data: {
-        getUser: {
+        user: {
           name: 'Arlene L McMahon'
         },
-        user: {
+        User: {
           name: 'William B Ropp'
         }
       }
