@@ -5,12 +5,18 @@
 
 'use strict'
 
+import { beforeAll, expect, test } from '@jest/globals'
+import { readFileSync } from 'fs'
 import { graphql, GraphQLSchema } from 'graphql'
-import { afterAll, beforeAll, expect, test } from '@jest/globals'
+import { join } from 'path'
 
 import * as openAPIToGraphQL from '../src/index'
 
-const oas = require('./fixtures/example_oas4.json')
+function getOas() {
+  const oasStr = readFileSync(join(__dirname, './fixtures/example_oas4.json'), 'utf8');
+  const oas = JSON.parse(oasStr);
+  return oas;
+};
 
 let createdSchema: GraphQLSchema
 
@@ -19,7 +25,7 @@ let createdSchema: GraphQLSchema
 // Set up the schema
 beforeAll(() => {
   return openAPIToGraphQL
-    .createGraphQLSchema(oas)
+    .createGraphQLSchema(getOas())
     .then(({ schema, report }) => {
       createdSchema = schema
     })
@@ -77,18 +83,18 @@ const oneOfQuery = `{
  * Because they are the same, the created GraphQL object should only have one
  * 'commonAttribute' field
  */
-test('Basic anyOf test using the same member schemas\n\nEquivalent to GET /anyOf', () => {
-  return graphql(createdSchema, anyOfQuery).then((result) => {
+test('Basic anyOf test using the same member schemas', () => {
+  return graphql({schema: createdSchema, source: anyOfQuery}).then((result) => {
     expect(
-      result.data['__schema'].queryType.fields.find((field) => {
+      (result.data['__schema'] as any).queryType.fields.find((field) => {
         return field.name === 'anyOf'
       })
     ).toEqual({
       name: 'anyOf',
       description:
-        'Basic anyOf test using the same member schemas\n\nEquivalent to GET /anyOf',
+        'Basic anyOf test using the same member schemas',
       type: {
-        name: 'AnyOf',
+        name: 'anyOf_200_response',
         kind: 'OBJECT',
         fields: [
           {
@@ -112,18 +118,18 @@ test('Basic anyOf test using the same member schemas\n\nEquivalent to GET /anyOf
  * Because they are the different, the created GraphQL object should have both
  * fields
  */
-test('Basic anyOf test with different member schemas\n\nEquivalent to GET /anyOf2', () => {
-  return graphql(createdSchema, anyOfQuery).then((result) => {
+test('Basic anyOf test with different member schemas', () => {
+  return graphql({schema: createdSchema, source: anyOfQuery}).then((result) => {
     expect(
-      result.data['__schema'].queryType.fields.find((field) => {
+      (result.data['__schema'] as any).queryType.fields.find((field) => {
         return field.name === 'anyOf2'
       })
     ).toEqual({
       name: 'anyOf2',
       description:
-        'Basic anyOf test with different member schemas\n\nEquivalent to GET /anyOf2',
+        'Basic anyOf test with different member schemas',
       type: {
-        name: 'AnyOf2',
+        name: 'anyOf2_200_response',
         kind: 'OBJECT',
         fields: [
           {
@@ -152,24 +158,24 @@ test('Basic anyOf test with different member schemas\n\nEquivalent to GET /anyOf
  * Because they are the same, the created GraphQL object should only have one
  * field
  */
-test('anyOf test with the same nested member schemas\n\nEquivalent to GET /anyOf3', () => {
-  return graphql(createdSchema, anyOfQuery).then((result) => {
+test('anyOf test with the same nested member schemas', () => {
+  return graphql({schema: createdSchema, source: anyOfQuery}).then((result) => {
     expect(
-      result.data['__schema'].queryType.fields.find((field) => {
+      (result.data['__schema'] as any).queryType.fields.find((field) => {
         return field.name === 'anyOf3'
       })
     ).toEqual({
       name: 'anyOf3',
       description:
-        'anyOf test with the same nested member schemas\n\nEquivalent to GET /anyOf3',
+        'anyOf test with the same nested member schemas',
       type: {
-        name: 'AnyOf3',
+        name: 'anyOf3_200_response',
         kind: 'OBJECT',
         fields: [
           {
             name: 'commonAttribute',
             type: {
-              name: 'CommonAttribute'
+              name: 'commonAttributeObject'
             }
           }
         ]
@@ -187,30 +193,30 @@ test('anyOf test with the same nested member schemas\n\nEquivalent to GET /anyOf
  * Because they are different at the root level, the created GraphQL object
  * should have two fields.
  */
-test('anyOf test with different nested member schemas\n\nEquivalent to GET /anyOf4', () => {
-  return graphql(createdSchema, anyOfQuery).then((result) => {
+test('anyOf test with different nested member schemas', () => {
+  return graphql({schema: createdSchema, source: anyOfQuery}).then((result) => {
     expect(
-      result.data['__schema'].queryType.fields.find((field) => {
+      (result.data['__schema'] as any).queryType.fields.find((field) => {
         return field.name === 'anyOf4'
       })
     ).toEqual({
       name: 'anyOf4',
       description:
-        'anyOf test with different nested member schemas\n\nEquivalent to GET /anyOf4',
+        'anyOf test with different nested member schemas',
       type: {
-        name: 'AnyOf4',
+        name: 'anyOf4_200_response',
         kind: 'OBJECT',
         fields: [
           {
             name: 'commonAttribute',
             type: {
-              name: 'CommonAttribute'
+              name: 'commonAttributeObject'
             }
           },
           {
             name: 'differentAttribute',
             type: {
-              name: 'DifferentAttribute'
+              name: 'commonAttributeObject'
             }
           }
         ]
@@ -228,18 +234,18 @@ test('anyOf test with different nested member schemas\n\nEquivalent to GET /anyO
  * This leads to a conlict because the same field has different schemas. As a
  * result, the field will use the arbitrary JSON type.
  */
-test('anyOf test with different nested member schemas, leading to conflict\n\nEquivalent to GET /anyOf5', () => {
-  return graphql(createdSchema, anyOfQuery).then((result) => {
+test('anyOf test with different nested member schemas, leading to conflict', () => {
+  return graphql({schema: createdSchema, source: anyOfQuery}).then((result) => {
     expect(
-      result.data['__schema'].queryType.fields.find((field) => {
+      (result.data['__schema'] as any).queryType.fields.find((field) => {
         return field.name === 'anyOf5'
       })
     ).toEqual({
       name: 'anyOf5',
       description:
-        'anyOf test with different nested member schemas, leading to conflict\n\nEquivalent to GET /anyOf5',
+        'anyOf test with different nested member schemas, leading to conflict',
       type: {
-        name: 'AnyOf5',
+        name: 'anyOf5_200_response',
         kind: 'OBJECT',
         fields: [
           {
@@ -263,20 +269,33 @@ test('anyOf test with different nested member schemas, leading to conflict\n\nEq
  * This leads to a conlict. As a result, the field will use the arbitrary JSON
  * type.
  */
-test('anyOf test with incompatible member schema types\n\nEquivalent to GET /anyOf6', () => {
-  return graphql(createdSchema, anyOfQuery).then((result) => {
+test('anyOf test with incompatible member schema types', () => {
+  return graphql({schema: createdSchema, source: anyOfQuery}).then((result) => {
     expect(
-      result.data['__schema'].queryType.fields.find((field) => {
+      (result.data['__schema'] as any).queryType.fields.find((field) => {
         return field.name === 'anyOf6'
       })
     ).toEqual({
       name: 'anyOf6',
       description:
-        'anyOf test with incompatible member schema types\n\nEquivalent to GET /anyOf6',
+        'anyOf test with incompatible member schema types',
       type: {
-        name: 'JSON',
-        kind: 'SCALAR',
-        fields: null
+        fields: [
+          {
+            name: "commonAttribute",
+            type: {
+              name: "String",
+            },
+          },
+          {
+            name: "String",
+            type: {
+              name: "String",
+            },
+          },
+        ],
+        kind: "OBJECT",
+        name: "anyOf6_200_response",
       }
     })
   })
@@ -290,24 +309,30 @@ test('anyOf test with incompatible member schema types\n\nEquivalent to GET /any
  * The created type should be able to pick out the object type schema without
  * defaulting to the arbitrary JSON type.
  */
-test('anyOf test with some extraneous member schemas\n\nEquivalent to GET /anyOf7', () => {
-  return graphql(createdSchema, anyOfQuery).then((result) => {
+test('anyOf test with some extraneous member schemas', () => {
+  return graphql({schema: createdSchema, source: anyOfQuery}).then((result) => {
     expect(
-      result.data['__schema'].queryType.fields.find((field) => {
+      (result.data['__schema'] as any).queryType.fields.find((field) => {
         return field.name === 'anyOf7'
       })
     ).toEqual({
       name: 'anyOf7',
       description:
-        'anyOf test with some extraneous member schemas\n\nEquivalent to GET /anyOf7',
+        'anyOf test with some extraneous member schemas',
       type: {
-        name: 'AnyOf7',
+        name: 'anyOf7_200_response',
         kind: 'OBJECT',
         fields: [
           {
             name: 'commonAttribute',
             type: {
               name: 'String'
+            }
+          },
+          {
+            name: 'Float',
+            type: {
+              name: 'Float'
             }
           }
         ]
@@ -323,20 +348,33 @@ test('anyOf test with some extraneous member schemas\n\nEquivalent to GET /anyOf
  * target type and the other two have no target types. Therefore, use integer
  * type.
  */
-test('anyOf test with no object type member schemas\n\nEquivalent to GET /anyOf8', () => {
-  return graphql(createdSchema, anyOfQuery).then((result) => {
+test('anyOf test with no object type member schemas', () => {
+  return graphql({schema: createdSchema, source: anyOfQuery}).then((result) => {
     expect(
-      result.data['__schema'].queryType.fields.find((field) => {
+      (result.data['__schema'] as any).queryType.fields.find((field) => {
         return field.name === 'anyOf8'
       })
     ).toEqual({
       name: 'anyOf8',
       description:
-        'anyOf test with no object type member schemas\n\nEquivalent to GET /anyOf8',
+        'anyOf test with no object type member schemas',
       type: {
-        name: 'Int',
-        kind: 'SCALAR',
-        fields: null
+        name: 'anyOf8_200_response',
+        kind: 'OBJECT',
+        fields: [
+          {
+            name: "Int",
+            type: {
+              name: "Int",
+            }
+          },
+          {
+            name: "Float",
+            type: {
+              name: "Float",
+            }
+          }
+        ]
       }
     })
   })
@@ -348,16 +386,16 @@ test('anyOf test with no object type member schemas\n\nEquivalent to GET /anyOf8
  * None of the member schemas are object type schemas but because there is an
  * external type provided in the root schema, it can utilize the proper typing.
  */
-test('anyOf test with extraneous member schemas with external type\n\nEquivalent to GET /anyOf9', () => {
-  return graphql(createdSchema, anyOfQuery).then((result) => {
+test('anyOf test with extraneous member schemas with external type', () => {
+  return graphql({schema: createdSchema, source: anyOfQuery}).then((result) => {
     expect(
-      result.data['__schema'].queryType.fields.find((field) => {
+      (result.data['__schema'] as any).queryType.fields.find((field) => {
         return field.name === 'anyOf9'
       })
     ).toEqual({
       name: 'anyOf9',
       description:
-        'anyOf test with extraneous member schemas with external type\n\nEquivalent to GET /anyOf9',
+        'anyOf test with extraneous member schemas with external type',
       type: {
         name: 'Int',
         kind: 'SCALAR',
@@ -372,17 +410,17 @@ test('anyOf test with extraneous member schemas with external type\n\nEquivalent
  *
  * None of the schemas have conflicts so all three should be utilized
  */
-test('Basic anyOf test with allOf\n\nEquivalent to GET /anyOf10', () => {
-  return graphql(createdSchema, anyOfQuery).then((result) => {
+test('Basic anyOf test with allOf', () => {
+  return graphql({schema: createdSchema, source: anyOfQuery}).then((result) => {
     expect(
-      result.data['__schema'].queryType.fields.find((field) => {
+      (result.data['__schema'] as any).queryType.fields.find((field) => {
         return field.name === 'anyOf10'
       })
     ).toEqual({
       name: 'anyOf10',
-      description: 'Basic anyOf test with allOf\n\nEquivalent to GET /anyOf10',
+      description: 'Basic anyOf test with allOf',
       type: {
-        name: 'AnyOf10',
+        name: 'anyOf10_200_response',
         kind: 'OBJECT',
         fields: [
           {
@@ -416,18 +454,18 @@ test('Basic anyOf test with allOf\n\nEquivalent to GET /anyOf10', () => {
  * Resolving the allOf should correctly collapse all of the (nested) anyOfs
  * and allow all three schemas to be utilized
  */
-test('anyOf test with allOf, requiring anyOf collapse\n\nEquivalent to GET /anyOf11', () => {
-  return graphql(createdSchema, anyOfQuery).then((result) => {
+test('anyOf test with allOf, requiring anyOf collapse', () => {
+  return graphql({schema: createdSchema, source: anyOfQuery}).then((result) => {
     expect(
-      result.data['__schema'].queryType.fields.find((field) => {
+      (result.data['__schema'] as any).queryType.fields.find((field) => {
         return field.name === 'anyOf11'
       })
     ).toEqual({
       name: 'anyOf11',
       description:
-        'anyOf test with allOf, requiring anyOf collapse\n\nEquivalent to GET /anyOf11',
+        'anyOf test with allOf, requiring anyOf collapse',
       type: {
-        name: 'AnyOf11',
+        name: 'anyOf11_200_response',
         kind: 'OBJECT',
         fields: [
           {
@@ -460,21 +498,21 @@ test('anyOf test with allOf, requiring anyOf collapse\n\nEquivalent to GET /anyO
  * Because the schemas are different object types, the created GraphQL union
  * type has two differnet member types.
  */
-test('Basic oneOf test\n\nEquivalent to GET /oneOf', () => {
-  return graphql(createdSchema, oneOfQuery).then((result) => {
+test('Basic oneOf test', () => {
+  return graphql({schema: createdSchema, source: oneOfQuery}).then((result) => {
     expect(
-      result.data['__schema'].queryType.fields.find((field) => {
+      (result.data['__schema'] as any).queryType.fields.find((field) => {
         return field.name === 'oneOf'
       })
     ).toEqual({
       name: 'oneOf',
-      description: 'Basic oneOf test\n\nEquivalent to GET /oneOf',
+      description: 'Basic oneOf test',
       type: {
-        name: 'OneOf',
+        name: 'oneOf_200_response',
         kind: 'UNION',
         possibleTypes: [
           {
-            name: 'CommonAttributeObject',
+            name: 'commonAttributeObject',
             fields: [
               {
                 type: {
@@ -484,7 +522,7 @@ test('Basic oneOf test\n\nEquivalent to GET /oneOf', () => {
             ]
           },
           {
-            name: 'DifferentAttributeObject',
+            name: 'differentAttributeObject',
             fields: [
               {
                 type: {
@@ -505,20 +543,41 @@ test('Basic oneOf test\n\nEquivalent to GET /oneOf', () => {
  * Because one of the member schemas is not an object type, then default to
  * the arbitrary JSON type.
  */
-test('oneOf test with non-object type member schema\n\nEquivalent to GET /oneOf2', () => {
-  return graphql(createdSchema, oneOfQuery).then((result) => {
+test('oneOf test with non-object type member schema', () => {
+  return graphql({schema: createdSchema, source: oneOfQuery}).then((result) => {
     expect(
-      result.data['__schema'].queryType.fields.find((field) => {
+      (result.data['__schema'] as any).queryType.fields.find((field) => {
         return field.name === 'oneOf2'
       })
     ).toEqual({
       name: 'oneOf2',
       description:
-        'oneOf test with non-object type member schema\n\nEquivalent to GET /oneOf2',
+        'oneOf test with non-object type member schema',
       type: {
-        name: 'JSON',
-        kind: 'SCALAR',
-        possibleTypes: null
+        name: "oneOf2_200_response",
+        kind: "UNION",
+        possibleTypes: [
+          {
+            fields: [
+              {
+                type: {
+                  name: "String",
+                },
+              },
+            ],
+            name: "commonAttributeObject",
+          },
+          {
+            fields: [
+              {
+                type: {
+                  name: "Int",
+                },
+              },
+            ],
+            name: "Int_container",
+          },
+        ],
       }
     })
   })
@@ -530,20 +589,41 @@ test('oneOf test with non-object type member schema\n\nEquivalent to GET /oneOf2
  * None of the member schemas are object types, therefore default to
  * the arbitrary JSON type.
  */
-test('oneOf test with no object type member schemas\n\nEquivalent to GET /oneOf3', () => {
-  return graphql(createdSchema, oneOfQuery).then((result) => {
+test('oneOf test with no object type member schemas', () => {
+  return graphql({schema: createdSchema, source: oneOfQuery}).then((result) => {
     expect(
-      result.data['__schema'].queryType.fields.find((field) => {
+      (result.data['__schema'] as any).queryType.fields.find((field) => {
         return field.name === 'oneOf3'
       })
     ).toEqual({
       name: 'oneOf3',
       description:
-        'oneOf test with no object type member schemas\n\nEquivalent to GET /oneOf3',
+        'oneOf test with no object type member schemas',
       type: {
-        name: 'JSON',
-        kind: 'SCALAR',
-        possibleTypes: null
+        name: 'oneOf3_200_response',
+        kind: 'UNION',
+        possibleTypes: [
+          {
+            name: "String_container",
+            fields: [
+              {
+                type: {
+                  name: 'String'
+                }
+              }
+            ]
+          },
+          {
+            name: "Int_container",
+            fields: [
+              {
+                type: {
+                  name: "Int"
+                }
+              }
+            ]
+          }
+        ]
       }
     })
   })
@@ -555,16 +635,16 @@ test('oneOf test with no object type member schemas\n\nEquivalent to GET /oneOf3
  * The member schemas contain extranous data but because the root schema contains a
  * type, it is able to utilize the proper type.
  */
-test('oneOf test with extraneous member schemas\n\nEquivalent to GET /oneOf4', () => {
-  return graphql(createdSchema, oneOfQuery).then((result) => {
+test('oneOf test with extraneous member schemas', () => {
+  return graphql({schema: createdSchema, source: oneOfQuery}).then((result) => {
     expect(
-      result.data['__schema'].queryType.fields.find((field) => {
+      (result.data['__schema'] as any).queryType.fields.find((field) => {
         return field.name === 'oneOf4'
       })
     ).toEqual({
       name: 'oneOf4',
       description:
-        'oneOf test with extraneous member schemas\n\nEquivalent to GET /oneOf4',
+        'oneOf test with extraneous member schemas',
       type: {
         name: 'Int',
         kind: 'SCALAR',
@@ -581,21 +661,21 @@ test('oneOf test with extraneous member schemas\n\nEquivalent to GET /oneOf4', (
  *
  * TODO: verify this behavior and also create a test with additional root properties
  */
-test('Basic oneOf test with allOf\n\nEquivalent to GET /oneOf5', () => {
-  return graphql(createdSchema, oneOfQuery).then((result) => {
+test('Basic oneOf test with allOf', () => {
+  return graphql({schema: createdSchema, source: oneOfQuery}).then((result) => {
     expect(
-      result.data['__schema'].queryType.fields.find((field) => {
+      (result.data['__schema'] as any).queryType.fields.find((field) => {
         return field.name === 'oneOf5'
       })
     ).toEqual({
       name: 'oneOf5',
-      description: 'Basic oneOf test with allOf\n\nEquivalent to GET /oneOf5',
+      description: 'Basic oneOf test with allOf',
       type: {
-        name: 'OneOf5',
+        name: 'oneOf5_200_response',
         kind: 'UNION',
         possibleTypes: [
           {
-            name: 'CommonAttributeObject',
+            name: 'commonAttributeObject',
             fields: [
               {
                 type: {
@@ -605,7 +685,7 @@ test('Basic oneOf test with allOf\n\nEquivalent to GET /oneOf5', () => {
             ]
           },
           {
-            name: 'DifferentAttributeObject',
+            name: 'differentAttributeObject',
             fields: [
               {
                 type: {
@@ -627,22 +707,22 @@ test('Basic oneOf test with allOf\n\nEquivalent to GET /oneOf5', () => {
  * Resolving the allOf should correctly collapse all of the (nested) oneOfs
  * and allow all three schemas to be utilized
  */
-test('oneOf test with allOf, requiring oneOf collapse\n\nEquivalent to GET /oneOf6', () => {
-  return graphql(createdSchema, oneOfQuery).then((result) => {
+test('oneOf test with allOf, requiring oneOf collapse', () => {
+  return graphql({schema: createdSchema, source: oneOfQuery}).then((result) => {
     expect(
-      result.data['__schema'].queryType.fields.find((field) => {
+      (result.data['__schema'] as any).queryType.fields.find((field) => {
         return field.name === 'oneOf6'
       })
     ).toEqual({
       name: 'oneOf6',
       description:
-        'oneOf test with allOf, requiring oneOf collapse\n\nEquivalent to GET /oneOf6',
+        'oneOf test with allOf, requiring oneOf collapse',
       type: {
-        name: 'OneOf6',
+        name: 'oneOf6_200_response',
         kind: 'UNION',
         possibleTypes: [
           {
-            name: 'CommonAttributeObject',
+            name: 'commonAttributeObject',
             fields: [
               {
                 type: {
@@ -652,7 +732,7 @@ test('oneOf test with allOf, requiring oneOf collapse\n\nEquivalent to GET /oneO
             ]
           },
           {
-            name: 'DifferentAttributeObject',
+            name: 'differentAttributeObject',
             fields: [
               {
                 type: {
@@ -662,7 +742,7 @@ test('oneOf test with allOf, requiring oneOf collapse\n\nEquivalent to GET /oneO
             ]
           },
           {
-            name: 'AnotherAttributeObject',
+            name: 'anotherAttributeObject',
             fields: [
               {
                 type: {
@@ -685,15 +765,15 @@ test('oneOf test with allOf, requiring oneOf collapse\n\nEquivalent to GET /oneO
  * Resolving the oneOf and allOfs should correctly create a union of two object
  * types, each object type with a link field from the oneOf schema
  */
- test('oneOf test with allOfs, requiring oneOf collapse\n\nEquivalent to GET /OneOfWithAllOfsAndLink', () => {
-  return graphql(createdSchema, oneOfQuery).then((result) => {
+ test('oneOf test with allOfs, requiring oneOf collapse', () => {
+  return graphql({schema: createdSchema, source: oneOfQuery}).then((result) => {
     expect(
-      result.data['__schema'].queryType.fields.find((field) => {
+      (result.data['__schema'] as any).queryType.fields.find((field) => {
         return field.name === 'oneOfWithAllOfsAndLink'
       })
     ).toEqual({
       "name": "oneOfWithAllOfsAndLink",
-      "description": "Equivalent to GET /oneOfWithAllOfsAndLink",
+      "description": null,
       "type": {
         "name": "OneOfWithAllOfsAndLink",
         "kind": "UNION",
@@ -721,11 +801,6 @@ test('oneOf test with allOf, requiring oneOf collapse\n\nEquivalent to GET /oneO
           {
             "name": "Two",
             "fields": [
-              {
-                "type": {
-                  "name": "String"
-                }
-              },
               {
                 "type": {
                   "name": "String"
